Update index.php
Last time my solution of using  $pi->init(); solved the cold boot issue and it worked fine in the past 6 months for me since I only use INPUT,  but now I am trying to use the relay of PiFace  and the problem is that every time I run the script, init() resets all the ports. In other words, if my relay was set to ON, and whenever I call the rest API again, the init() resets everything to OFF.  Now, I have a workaround although it looks a bit ugly but it works.